IT DISPLAYS IMAGES! OMGWTFBBQ
Major changes:
Slightly changed how a ComicStrip works. It can now take in either a string OR a URL file location. If given a URL, the loadImage method will work slightly differently, but otherwise everything works the same. The isOnline method can be used to see if it was given a URL or not.

Added error messages to ComicDialog and now forces users to enter comic names.

Changed MainFrame around a bunch so it actually DISPLAYS comics and their text. Check it out. It totally works.