MainFrame: Changed around MainFrame so that it uses a right-click dropdown box to interact with the current comic, rather than menu items at the top. Added a confirmation dialog to comic removal.
ComicDatabase:
Added a simple hasComics method to ComicDatabase (returns !allComics.isEmpty()) for use in MainFrame.

ComicDialog:
Fixed a small problem in which the default button when editing an existing ComicSite is the Edit URL button instead of Accept. Added two dialog methods to simplify error message code.

BaconSystem and AboutFrame:
Made the text in AboutFrame a constant at the top of BaconSystem (ABOUT_TEXT). The text can be put into an AboutFrame as is or displayed in a regular dialog box if preferrable (made aboutButton bring up both options for comparison).